Created Exceptions Handler
added an exceptions_handler() function that works with both state change reactions and narratives. Both these can use general/exceptions sets, so I used the same syntax in the YAML file and was able to refactor the code to only have to write the code once, in the aforementioned exceptions_handler() function.